MNES-1283: Sorting files renamed
